Compile all source files

+++ code/midtown/mmaudio/audstream.cpp
@@ -20,5 +20,5 @@ define_dummy_symbol(mmaudio_audstream);
 
 #include "audstream.h"
 
-void AudStream::SetPlayPosition(u32 /*arg1*/)
+void AudStream::SetPlayPosition(ulong /*arg1*/)
 {}
